+++ NEWS.md
@@ -3,3 +3,4 @@
 * Added a `NEWS.md` file to track changes to the package.
 * Added a `README.Rmd` file.
 * Added Github Repo.
+* Added the function `theme_du_bois` and dependency `ggplot2`